Fix bug with temp file location
Previously could not save the calendar file
when installed in Program Files. This is because
a bug was causing the path for temp files
to be "./", which was evaluating to the program
folder in Program Files, which the program
obviously does not have permissions to write to.
The program now correctly stores the temp file
in the same directory as the calendar file.