Simplify tcl/tk detection by a comical amount
Yet to see if this detection is as... "resilient" as the old stuff.
Clearly it is on modern Linux because it can find my tcl/tk install where the old code could not,
but that might not hold true for Richard's CentOS install.
The configure command in his install instructions has tcl placed differently.
Gotta ask him to test it.