Refactor "delete all before" a lot
* Now the ItemListing has a function that just shows all items in a given list;
much better than the previous hacky solution for hiding "important" items of
just adding a flag that doesnt show them. Much more generic.
* `ical_deleteallbefore` has been split into three parts: one that prompts for the date, one that confirms the user wants to delete the chosen items,
and one that deletes the items. much more generic:tm:.
* It can now handle repeating items "properly". It would be optimal if deleting a chunk of days off a repeating item also made a repeating slice in the delete history,
but that's just unnecessary complication. For now it just makes a bunch of non-repeating items in the delete history.