Implement basic autopurge
Pending a switch that lets the user choose if it asks or not. For now itll always ask, cause thats safer.
Also, fixed a relevant bug in my Time refactor that broke constructing a Time object from a double.
I was performing `int(r * 1000000)` which caused overflows every time, since r is the time since unix epoch in seconds.
Fix is to cast to (long long) instead, which is the datatype that the chrono::microseconds type uses.